Split call_typedoc into main.ts and cli.ts. NFC
This moves most of the logic from call_typedoc.ts into a separate file called
`cli.ts` which avoids some of the io like `process.exit()` and serializing
json / writing to a file. Intended to help with testing.
